@@ -1,36 +1,30 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace UID2.Client
 {
-    public struct DecryptionResponse
+    public readonly struct DecryptionResponse
     {
-        private readonly DecryptionStatus _status;
-        private readonly string _uid;
-        private readonly DateTime? _established;
-        private readonly int? _siteId;
-        private readonly int? _siteKeySiteId;
-
-        public DecryptionResponse(DecryptionStatus status, string uid, DateTime? established, int? siteId, int? siteKeySiteId)
+        public DecryptionResponse(DecryptionStatus status, string uid, DateTime? established, int? siteId, int? siteKeySiteId, bool? isClientSideGenerated = false)
         {
-            _status = status;
-            _uid = uid;
-            _established = established;
-            _siteId = siteId;
-            _siteKeySiteId = siteKeySiteId;
+            Status = status;
+            Uid = uid;
+            Established = established;
+            SiteId = siteId;
+            SiteKeySiteId = siteKeySiteId;
+            IsClientSideGenerated = isClientSideGenerated;
         }
 
         public static DecryptionResponse MakeError(DecryptionStatus status)
         {
-            return new DecryptionResponse(status, null, null, null, null);
+            return new DecryptionResponse(status, null, null, null, null, null);
         }
 
-        public bool Success => _status == DecryptionStatus.Success;
-        public DecryptionStatus Status => _status;
-        public string Uid => _uid;
-        public DateTime? Established => _established;
-        public int? SiteId => _siteId;
-        public int? SiteKeySiteId => _siteKeySiteId;
+        public bool Success => Status == DecryptionStatus.Success;
+        public DecryptionStatus Status { get; }
+        public string Uid { get; }
+        public DateTime? Established { get; }
+        public int? SiteId { get; }
+        public int? SiteKeySiteId { get; }
+        public bool? IsClientSideGenerated { get; }
     }
 }
